+++ app/src/main/java/com/example/randomuserapp/repository/UserRepository.kt
@@ -1,34 +1,7 @@
 package com.example.randomuserapp.repository
 
-import android.util.Log
-import com.example.randomuserapp.data.api.RetrofitInstance
-import com.example.randomuserapp.data.db.UserDao
 import com.example.randomuserapp.data.db.UserEntity
 
-class UserRepository(private val userDao: UserDao) : UserRepositoryContract {
-
-    override suspend fun getUsers(): List<UserEntity> {
-        return try {
-            val response = RetrofitInstance.api.getUsers()
-            val users = response.results.map {
-                UserEntity(
-                    firstName = it.name.first,
-                    lastName = it.name.last,
-                    email = it.email,
-                    avatarUrl = it.picture.large
-                )
-            }
-
-            userDao.deleteAllUsers()
-            userDao.insertUsers(users)
-
-            users
-        } catch (e: Exception) {
-            Log.e("UserRepository", "Ошибка при получении пользователей", e)
-            userDao.getAllUsers()
-        }
-    }
+interface UserRepository {
+    suspend fun getUsers(): List<UserEntity>
 }
-
-
-
