--- app/src/main/java/com/example/randomuserapp/repository/UserRepositoryContract.kt
@@ -1,7 +1,0 @@
-package com.example.randomuserapp.repository
-
-import com.example.randomuserapp.data.db.UserEntity
-
-interface UserRepositoryContract {
-    suspend fun getUsers(): List<UserEntity>
-}
