+++ app/src/main/java/com/example/randomuserapp/viewmodel/UserViewModel.kt
@@ -3,7 +3,7 @@ package com.example.randomuserapp.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.randomuserapp.data.db.UserEntity
-import com.example.randomuserapp.repository.UserRepository
+import com.example.randomuserapp.repository.UserRepositoryImpl
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toPersistentList
@@ -13,7 +13,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-class UserViewModel(private val repository: UserRepository) : ViewModel() {
+class UserViewModel(private val repository: UserRepositoryImpl) : ViewModel() {
 
     private val _userList = MutableStateFlow<PersistentList<UserEntity>>(persistentListOf())
     val userList = _userList.asStateFlow()
