+++ app/src/main/java/com/example/randomuserapp/viewmodel/UserViewModelFactory.kt
@@ -2,9 +2,9 @@ package com.example.randomuserapp.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.randomuserapp.repository.UserRepository
+import com.example.randomuserapp.repository.UserRepositoryImpl
 
-class UserViewModelFactory(private val repository: UserRepository) : ViewModelProvider.Factory {
+class UserViewModelFactory(private val repository: UserRepositoryImpl) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(UserViewModel::class.java)) {
             @Suppress("UNCHECKED_CAST")
